+++ src/ref-parser/ref-parser.js
@@ -2,7 +2,11 @@ import { mergeDeep } from '../helpers/utils.js'
 import JsonWalker from '../json-walker.js'
 
 class RefParser {
-  constructor () {
+  constructor (options = {}) {
+    this.options = Object.assign({
+      detectRecursion: true
+    }, options)
+
     this.refs = {}
     this.data = {}
     this.iterations = 0
@@ -29,8 +33,10 @@ class RefParser {
       console.warn('Missing refs:', JSON.stringify(missingRefs))
     }
 
-    this.cycles = this.findRecursiveRefs(this.refs)
-    this.markRecursiveSchemas()
+    if (this.options.detectRecursion) {
+      this.cycles = this.findRecursiveRefs(this.refs)
+      this.markRecursiveSchemas()
+    }
   }
 
   refsResolved () {
@@ -134,7 +140,7 @@ class RefParser {
   }
 
   hasRefCycles () {
-    return this.cycles.length > 0
+    return this.options.detectRecursion && this.cycles.length > 0
   }
 
   markRecursiveSchemas () {
@@ -146,7 +152,7 @@ class RefParser {
     })
 
     // Mark schemas with $refs pointing to cycle members
-    for (const [path, schema] of Object.entries(this.data)) {
+    for (const schema of Object.values(this.data)) {
       if (schema && schema.$ref && cycleRefs.has(schema.$ref)) {
         schema['x-recursive'] = true
       }